telegram_bot_railway/requirements.txt
python-telegram-bot==13.15
google-api-python-client==2.120.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dateutil==2.9.0.post0